loadbalancers: moving the existing data sources/resources over
