+++ azurerm/internal/services/loadbalancer/lb_backend_address_pool_data_source.go
@@ -1,4 +1,4 @@
-package network
+package loadbalancer
 
 import (
 	"fmt"
