+++ azurerm/internal/services/loadbalancer/lb_nat_pool_resource.go
@@ -1,4 +1,4 @@
-package network
+package loadbalancer
 
 import (
 	"fmt"
