+++ azurerm/internal/services/loadbalancer/lb_nat_rule_resource.go
@@ -1,4 +1,4 @@
-package network
+package loadbalancer
 
 import (
 	"fmt"
