+++ azurerm/internal/services/loadbalancer/loadbalancer.go
@@ -1,4 +1,4 @@
-package network
+package loadbalancer
 
 import (
 	"context"
