+++ azurerm/internal/services/loadbalancer/tests/loadbalancer_nat_pool_resource_test.go
