+++ azurerm/internal/services/loadbalancer/tests/loadbalancer_nat_rule_resource_test.go
